requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
python-zillow==0.2.0
requests==2.18.4
urllib3==1.22
xmltodict==0.11.0
